+++ packages/caption/__tests__/shared.js
@@ -7,7 +7,7 @@ import Caption from "../src/caption";
 const captionText = "Some caption text goes in here";
 const credits = "Just credits";
 const exampleImage =
-  "https://www.thetimes.co.uk/imageserver/image/methode%2Ftimes%2Fprod%2Fweb%2Fbin%2F7d2fd06c-a460-11e7-8955-1ad2a9a7928d.jpg?crop=1500%2C844%2C0%2C78&resize=685";
+  "https://www.thetimes.co.uk/imageserver/image/methode%2Ftimes%2Fprod%2Fweb%2Fbin%2F7d2fd06c-a460-11e7-8955-1ad2a9a7928d.jpg?crop=1500%2C844%2C0%2C78";
 const style = {
   container: {
     backgroundColor: "red"