+++ packages/utils/src/screen.js
@@ -1,0 +1,7 @@
+import { PixelRatio } from "react-native";
+
+import { acceptedWidths, normaliseWidth, screenWidth } from "./screen.base.js";
+
+export { acceptedWidths, normaliseWidth, screenWidth };
+export const screenWidthInPixels = () =>
+  PixelRatio.getPixelSizeForLayoutSize(screenWidth());