+++ packages/utils/src/screen.web.js
@@ -1,0 +1,3 @@
+/* eslint-disable import/prefer-default-export */
+
+export { normaliseWidth, screenWidth } from "./screen.base.js";